refactor: move ecstore rebalance control impls
